Mainline selection to cherry pick Jerry's commit didn't cooperate so just manually redid his changes and retested, appears to be fine.
