+++ qrequest.c
@@ -34,9 +34,3 @@ void    QuenosBlockSelf (void) {
     KernelBlock();
 }
 
-/**
- * Application binary interface documentation says other_pid will be passed into r4
- */
-void    QuenosUnblock (int other_pid) {
-    KernelUnblock();
-}
