+++ quser.c
@@ -25,6 +25,9 @@ static  char    P2stack[USER_STACK_SIZE];
 /**
  * Application binary interface documentation says other_pid will be passed into r4
  */
+void    QuenosUnblock (int other_pid) {
+    KernelUnblock();
+}
 
 
 // You will need to produce a basic operational Nios II kernel that can handle a couple of simple processes
@@ -34,8 +37,7 @@ static	void    Process1 (void)
         for (;;)
         {
                 printString("a\n");
-                //QuenosBlockSelf ();
-				QuenosRelinquish ();
+                KernelBlock();
         }
 }
 
@@ -44,8 +46,8 @@ static	void    Process2 (void)
         for (;;)
         {
                 printString("b\n");
-				//QuenosUnblock (1);
-				QuenosRelinquish ();
+				QuenosUnblock (1);
+				KernelRelinquish ();
         }
 }
